Create README.md
A group of workers came up with an idea to save money for a general cause. This enabled them to generate capital for a business idea, buy assets and accomplish other financial needs. They used MS excel sheets to track which person deposited money, on a date and sum up the total later. Several other smaller groups copied the same idea, and this has now become a norm among corporates in Africa. We have been tasked to come up with a system that will track the records of these workers and help them track their savings.
A member can belong to one or more saving groups and is required to emit his or her contributions in the first week of the month. Late payment penalty is 5% levied on savings submitted late. Members can view their account statements and apply for a loan. 
A member is eligible to take a loan if he or she has saved with a group for more than 6 months. In addition, a member can borrow up to twice his savings and requires seconding from 2 other members whose total savings exceed the loan (Loan should not exceed the requestors savings plus the total savings of the seconders).
A member can also request to leave a group. A 3 months’ notice should be served to the account manager and he should have cleared all his dues. A member can also view a summary report of his accounts if he belongs to one or more groups
New members can fill application to be added to a group but require seconding from two members of that group they have applied to.

Each group has an account manager who is also a member of the group. The account manager can approve loans, add new members, generate summary reports and organize meetings.